panelin_backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
